requirements.txt
flask
google-generativeai
google-cloud-secret-manager
google-cloud-aiplatform>=1.38
markdown2